refactor(sync): extract bindings + clone + models from sync_service.py (tech-debt)
`sync_service.py` had grown to ~4033 lines -- ~2.7x the 1500-LOC ceiling for
`services/*.py` (CONTRIBUTING.md calls this file out by name). #432 added ~321
LOC without splitting; this is the deferred split.

No behavior change -- pure mechanical extraction, public `SyncService` API
identical:

- `_sync_models.py`: the 5 dataclasses (WritebackResult, CreatedConfig,
  VariableBindingResult, FlowBindingResult, LocalConfigHashes) + the
  VARIABLES/FLOW component-id constants. `sync_service` re-exports `CreatedConfig`
  (the only externally-imported name) for back-compat.
- `_sync_bindings.py`: the push-time link backfill -- Phase C (variable links)
  and Phase D (keboola.flow task configIds) -- as free functions taking the
  `SyncService` as first arg (for the 3 on-disk helpers they need). push() now
  calls `resolve_variable_bindings(self, ...)` / `resolve_flow_task_bindings(self, ...)`.
- `_sync_clone.py`: the `clone_project` orchestration. `SyncService.clone_project`
  is now a thin delegator (public method preserved).

`sync_service.py`: 4033 -> 3435 lines (-598). The free-function-taking-service
pattern keeps typing explicit (ty-clean via a TYPE_CHECKING `SyncService` import)
and makes the bindings/clone unit-testable in isolation. Further extractions
(writeback, storage-metadata/jobs/samples) are natural follow-ups in the same
pattern. make check green (4092 passed).
