+++ src/keboola_agent_cli/services/sync_service.py
@@ -12,7 +12,6 @@ import shutil
 import subprocess
 import threading
 from concurrent.futures import ThreadPoolExecutor, as_completed
-from dataclasses import dataclass, field
 from pathlib import Path
 from typing import Any
 
@@ -62,92 +61,12 @@ from ._encryption import (
     encrypt_secrets_in_config,
     find_plaintext_secret_keys,
 )
+from ._sync_bindings import resolve_flow_task_bindings, resolve_variable_bindings
+from ._sync_clone import clone_project as _clone_project_impl
+from ._sync_models import CreatedConfig, LocalConfigHashes, WritebackResult
 from .base import BaseService
 
 logger = logging.getLogger(__name__)
-
-# Sibling component that backs a transformation's variable links. A
-# transformation references it via ``configuration.variables_id`` (the config)
-# and ``configuration.variables_values_id`` (a row id).
-VARIABLES_COMPONENT_ID = "keboola.variables"
-
-# Conditional-flow component. A flow runs other configs via
-# ``configuration.tasks[].task.configId`` (job-type tasks); the Phase-D backfill
-# remaps those ids placeholder/source -> ULID after a fresh create (e.g. clone).
-FLOW_COMPONENT_ID = "keboola.flow"
-
-
-@dataclass
-class WritebackResult:
-    """Outcome of recording a freshly-created config in the manifest.
-
-    ``previous_id`` is the manifest entry's id **before** the placeholder ->
-    ULID overwrite (empty string when a brand-new entry was appended). The
-    create pass uses it to key ``created_id_map`` so row parents and
-    transformation variable links can be remapped placeholder -> ULID.
-    """
-
-    entry: ManifestConfiguration
-    previous_id: str
-
-
-@dataclass
-class CreatedConfig:
-    """A config created during a single ``push`` create pass.
-
-    Carries just enough to drive the Phase-C variable-link backfill: the
-    component id, the API-assigned ULID, and the on-disk directory holding
-    the (post-writeback) ``_config.yml``.
-    """
-
-    component_id: str
-    config_id: str
-    config_dir: Path
-
-
-@dataclass
-class VariableBindingResult:
-    """Outcome of the Phase-C variable-link backfill.
-
-    ``configs_rewritten`` counts transformations whose remote configuration +
-    local ``_configuration_extra`` were rebound to ULIDs (drives the
-    manifest-dirty flag). ``errors`` accumulates unresolved links so the push
-    envelope surfaces them instead of leaving a broken link silently.
-    """
-
-    errors: list[dict[str, str]] = field(default_factory=list)
-    configs_rewritten: int = 0
-
-
-@dataclass
-class FlowBindingResult:
-    """Outcome of the Phase-D flow-task-link backfill (#426).
-
-    ``configs_rewritten`` counts flows whose task ``configId``s were remapped to
-    ULIDs (drives the manifest-dirty flag); ``tasks_remapped`` is the total task
-    references rewritten; ``errors`` accumulates PUT failures so the push
-    envelope surfaces them.
-    """
-
-    errors: list[dict[str, str]] = field(default_factory=list)
-    configs_rewritten: int = 0
-    tasks_remapped: int = 0
-
-
-@dataclass
-class LocalConfigHashes:
-    """Hashes describing a config dir's on-disk state after a push.
-
-    ``file_hash`` is the ``_config.yml`` content hash, ``cfg_hash`` the
-    normalized config hash (see :func:`config_hash`), and ``extra_hashes``
-    maps each extracted code/companion file to its hash. Stored on the
-    manifest entry so the next ``sync diff`` recognises local == remote.
-    """
-
-    file_hash: str
-    cfg_hash: str
-    extra_hashes: dict[str, str] = field(default_factory=dict)
-
 
 # Companion files extracted alongside a config's ``_config.yml`` whose hashes
 # are tracked so ``sync diff`` notices edits to code/description files.
@@ -1731,7 +1650,8 @@ class SyncService(BaseService):
                     self._record_push_error(errors, change_type, component_id, config_id, exc)
 
             # ---- Phase C: variable-link backfill (KFR-03) ----------------
-            binding = self._resolve_variable_bindings(
+            binding = resolve_variable_bindings(
+                self,
                 client,
                 created_configs=created_configs,
                 created_id_map=created_id_map,
@@ -1748,7 +1668,8 @@ class SyncService(BaseService):
             # After variable links, remap keboola.flow task configIds that point
             # at configs created this push (golden/placeholder -> ULID). Reuses
             # created_id_map; a no-op when no flow was created.
-            flow_binding = self._resolve_flow_task_bindings(
+            flow_binding = resolve_flow_task_bindings(
+                self,
                 client,
                 created_configs=created_configs,
                 created_id_map=created_id_map,
@@ -1789,136 +1710,20 @@ class SyncService(BaseService):
     ) -> dict[str, Any]:
         """Clone a reference synced project into a fresh target project (#426).
 
-        Copies the reference tree at ``source`` into ``target_dir``, applies the
-        declarative ``overrides`` (``bucket_map``, ``variable_values``,
-        ``instance_rename``), re-points the manifest at ``target_alias``'s
-        project, and pushes -- so every config is CREATEd fresh and its flow
-        task / variable links are remapped reference->ULID by push Phase C/D.
-
-        Cloning into a fresh target needs no id surgery: the reference's config
-        ids do not exist in the target remote, so the diff classifies every
-        config as ``added`` and ``created_id_map`` (keyed by the reference id)
-        drives the link remaps.
-
-        Idempotent: re-running with an existing ``target_dir`` skips the copy +
-        overrides and just pushes, so a clean clone then reports ``no_changes`` /
-        ``created: 0``.
-
-        Args:
-            source: A reference synced project dir (must contain
-                ``.keboola/manifest.json``).
-            target_alias: The project alias to clone INTO (must be a fresh
-                project on the first clone).
-            target_dir: Where to materialise the clone.
-            overrides: Optional dict with ``bucket_map`` (old->new bucket id),
-                ``variable_values`` (var name->value), and ``instance_rename``
-                (old path prefix->new path prefix) keys.
-            dry_run: Apply overrides + report the diff without pushing.
-            branch_override: Optional target branch id.
-
-        Returns:
-            A dict matching ``CloneResult``: ``status``
-            (``cloned`` | ``no_changes`` | ``dry_run``), ``target_alias``,
-            ``target_dir``, ``created``, ``bucket_rewrites``,
-            ``variable_overrides``, ``renamed_instances``, ``flow_task_remaps``,
-            ``push`` (the underlying push result), ``errors``.
-
-        Raises:
-            ConfigError: ``source`` is not a synced project, ``target_alias`` is
-                unknown, or (first clone) the target project already contains the
-                reference's configs (not a fresh target).
+        Thin delegator to :func:`._sync_clone.clone_project`; see that function
+        (and ``sync-workflow.md``) for the full behavior -- copy + parameterize
+        (bucket_map / variable_values / instance_rename) + push, with Phase-C/D
+        link remap, a fresh-target guard, and idempotent re-runs.
         """
-        from urllib.parse import urlparse
-
-        from ..sync.clone import (
-            apply_bucket_map,
-            apply_instance_rename,
-            apply_variable_values,
-            copy_reference_tree,
-            repoint_manifest_project,
+        return _clone_project_impl(
+            self,
+            source,
+            target_alias,
+            target_dir,
+            overrides=overrides,
+            dry_run=dry_run,
+            branch_override=branch_override,
         )
-
-        overrides = overrides or {}
-        bucket_map = overrides.get("bucket_map") or {}
-        variable_values = overrides.get("variable_values") or {}
-        instance_rename = overrides.get("instance_rename") or {}
-        source_dir = Path(source)
-        target_path = Path(target_dir)
-
-        # Validate the reference is a synced project.
-        try:
-            load_manifest(source_dir)
-        except FileNotFoundError as exc:
-            raise ConfigError(
-                f"Source {source_dir} is not a synced project (no "
-                ".keboola/manifest.json). Run `kbagent sync pull` there first."
-            ) from exc
-
-        target_project = self.resolve_projects([target_alias])[target_alias]
-
-        bucket_rewrites = 0
-        variable_overrides = 0
-        renamed_instances = 0
-        already_cloned = target_path.exists()
-
-        if not already_cloned:
-            copy_reference_tree(source_dir, target_path)
-            manifest = load_manifest(target_path)
-            repoint_manifest_project(
-                manifest,
-                project_id=target_project.project_id or 0,
-                api_host=urlparse(target_project.stack_url).netloc,
-            )
-            bucket_rewrites = apply_bucket_map(target_path, manifest, bucket_map)
-            variable_overrides = apply_variable_values(target_path, manifest, variable_values)
-            renamed_instances = apply_instance_rename(target_path, manifest, instance_rename)
-            save_manifest(target_path, manifest)
-
-        override_counts = {
-            "bucket_rewrites": bucket_rewrites,
-            "variable_overrides": variable_overrides,
-            "renamed_instances": renamed_instances,
-        }
-
-        if dry_run:
-            diff_result = self.diff(target_alias, target_path, branch_override=branch_override)
-            return {
-                "status": "dry_run",
-                "target_alias": target_alias,
-                "target_dir": str(target_path),
-                "summary": diff_result["summary"],
-                **override_counts,
-            }
-
-        # Fresh-target guard: on a first clone every one of OUR configs must be a
-        # CREATE. A non-'added' change means a reference id already exists in the
-        # target -> not a fresh target; refuse rather than UPDATE a stranger's config.
-        if not already_cloned:
-            diff_result = self.diff(target_alias, target_path, branch_override=branch_override)
-            collisions = [
-                f"{c.get('component_id')}/{c.get('config_id')}"
-                for c in diff_result["changes"]
-                if c["change_type"] != "added"
-            ]
-            if collisions:
-                raise ConfigError(
-                    "Clone requires a fresh target project, but these configs already "
-                    f"exist there: {', '.join(collisions[:10])}. Use a new/empty target "
-                    "project, or remove the existing configs first."
-                )
-
-        push_result = self.push(target_alias, target_path, branch_override=branch_override)
-        status = "no_changes" if push_result.get("status") == "no_changes" else "cloned"
-        return {
-            "status": status,
-            "target_alias": target_alias,
-            "target_dir": str(target_path),
-            "created": push_result.get("created", 0),
-            "flow_task_remaps": push_result.get("flow_task_remaps", 0),
-            "push": push_result,
-            "errors": push_result.get("errors", []),
-            **override_counts,
-        }
 
     # Kept for test compatibility: tests exercise fail-closed encryption via
     # ``SyncService._encrypt_secrets_in_config(...)``. Production code uses
@@ -2315,409 +2120,6 @@ class SyncService(BaseService):
         # Write back: update local file with encrypted secrets.
         # Use pristine_data so blocks/code stay only in their code files.
         self._writeback_after_push(pristine_data, config_dir, config_id, configuration)
-
-    def _resolve_variable_bindings(
-        self,
-        client: Any,
-        *,
-        created_configs: list[CreatedConfig],
-        created_id_map: dict[tuple[str, str], str],
-        created_row_id_map: dict[str, str],
-        created_rows_by_parent: dict[str, list[str]],
-        manifest: Manifest,
-        branch_id: int | None,
-    ) -> VariableBindingResult:
-        """Rebind transformation -> variables links from placeholders to ULIDs.
-
-        On a fresh CREATE the transformation config is POSTed with its
-        ``_configuration_extra.variables_id`` / ``variables_values_id`` still
-        set to the externally-authored placeholder strings (``config_format``
-        merges ``_configuration_extra`` into the API body verbatim). This pass,
-        run after the variables config and its values row have been created,
-        resolves each placeholder to the ULID assigned during this push, PUTs
-        the corrected configuration body, then rewrites the local file and
-        refreshes the manifest hashes so a re-push is clean (KFR-03).
-
-        Resolution is a no-op when no ``keboola.variables`` config was created
-        this push (the already-bound / UPDATE path). When the exact placeholder
-        key misses but exactly one ``keboola.variables`` config was created this
-        push, it binds to that one with a warning; zero or ambiguous (>1)
-        matches accumulate an error rather than writing a broken link.
-        """
-        result = VariableBindingResult()
-
-        created_variables_ulids = [
-            ulid
-            for (component_id, _placeholder), ulid in created_id_map.items()
-            if component_id == VARIABLES_COMPONENT_ID
-        ]
-
-        for created in created_configs:
-            if created.component_id == VARIABLES_COMPONENT_ID:
-                continue  # the variables config itself never carries a link
-            local_data = self._read_config_file(created.config_dir)
-            if local_data is None:
-                continue
-            extra = local_data.get("_configuration_extra")
-            if not isinstance(extra, dict):
-                continue
-            vars_placeholder = extra.get("variables_id")
-            if not vars_placeholder or not isinstance(vars_placeholder, str):
-                continue
-            raw_vals = extra.get("variables_values_id")
-            vals_placeholder = raw_vals if isinstance(raw_vals, str) else ""
-
-            parent_ulid = self._resolve_variables_parent(
-                created=created,
-                vars_placeholder=vars_placeholder,
-                created_id_map=created_id_map,
-                created_variables_ulids=created_variables_ulids,
-                errors=result.errors,
-            )
-            if parent_ulid is None:
-                continue
-
-            row_ulid = self._resolve_variables_row(
-                created=created,
-                parent_ulid=parent_ulid,
-                vals_placeholder=vals_placeholder,
-                created_row_id_map=created_row_id_map,
-                created_rows_by_parent=created_rows_by_parent,
-                errors=result.errors,
-            )
-            # A missing-but-required values row already recorded an error.
-            if vals_placeholder and row_ulid is None:
-                continue
-
-            try:
-                self._apply_variable_binding(
-                    client,
-                    created=created,
-                    local_data=local_data,
-                    parent_ulid=parent_ulid,
-                    row_ulid=row_ulid,
-                    manifest=manifest,
-                    branch_id=branch_id,
-                )
-            except KeboolaApiError as exc:
-                result.errors.append(
-                    {
-                        "change_type": "variable_link",
-                        "error_code": ErrorCode.VARIABLE_LINK_UNRESOLVED,
-                        "component_id": created.component_id,
-                        "config_id": created.config_id,
-                        "message": str(exc),
-                    }
-                )
-                continue
-            result.configs_rewritten += 1
-
-        return result
-
-    def _resolve_variables_parent(
-        self,
-        *,
-        created: CreatedConfig,
-        vars_placeholder: str,
-        created_id_map: dict[tuple[str, str], str],
-        created_variables_ulids: list[str],
-        errors: list[dict[str, str]],
-    ) -> str | None:
-        """Resolve a transformation's ``variables_id`` placeholder to a ULID.
-
-        Returns the ULID, or ``None`` when there is nothing to backfill
-        (already-bound path) or the link is ambiguous (an error is appended).
-        """
-        parent_ulid = created_id_map.get((VARIABLES_COMPONENT_ID, vars_placeholder))
-        if parent_ulid is not None:
-            return parent_ulid
-        if not created_variables_ulids:
-            # No variables config created this push: the link is either already
-            # a ULID (UPDATE path) or points outside this push. Leave it.
-            return None
-        if len(created_variables_ulids) == 1:
-            parent_ulid = created_variables_ulids[0]
-            logger.warning(
-                "Transformation %s/%s variables_id placeholder %r did not match any "
-                "created variables config; binding to the single keboola.variables "
-                "config created this push (%s).",
-                created.component_id,
-                created.config_id,
-                vars_placeholder,
-                parent_ulid,
-            )
-            return parent_ulid
-        errors.append(
-            {
-                "change_type": "variable_link",
-                "error_code": ErrorCode.VARIABLE_LINK_UNRESOLVED,
-                "component_id": created.component_id,
-                "config_id": created.config_id,
-                "message": (
-                    f"Cannot resolve variables_id placeholder {vars_placeholder!r}: "
-                    f"{len(created_variables_ulids)} keboola.variables configs were "
-                    "created this push and none matched by placeholder. Refusing to "
-                    "write an ambiguous variables link."
-                ),
-            }
-        )
-        return None
-
-    def _resolve_variables_row(
-        self,
-        *,
-        created: CreatedConfig,
-        parent_ulid: str,
-        vals_placeholder: str,
-        created_row_id_map: dict[str, str],
-        created_rows_by_parent: dict[str, list[str]],
-        errors: list[dict[str, str]],
-    ) -> str | None:
-        """Resolve a transformation's ``variables_values_id`` placeholder.
-
-        Returns the row ULID, or ``None`` when no values row was created (the
-        link is then left unset) or the choice is ambiguous (an error is
-        appended only when ``vals_placeholder`` was actually requested).
-        """
-        if vals_placeholder:
-            mapped = created_row_id_map.get(vals_placeholder)
-            if mapped is not None:
-                return mapped
-        siblings = created_rows_by_parent.get(parent_ulid, [])
-        if len(siblings) == 1:
-            row_ulid = siblings[0]
-            if vals_placeholder:
-                logger.warning(
-                    "Transformation %s/%s variables_values_id placeholder %r did not "
-                    "match a created row; binding to the single row created under "
-                    "variables config %s.",
-                    created.component_id,
-                    created.config_id,
-                    vals_placeholder,
-                    parent_ulid,
-                )
-            return row_ulid
-        if vals_placeholder:
-            errors.append(
-                {
-                    "change_type": "variable_link",
-                    "error_code": ErrorCode.VARIABLE_LINK_UNRESOLVED,
-                    "component_id": created.component_id,
-                    "config_id": created.config_id,
-                    "message": (
-                        f"Cannot resolve variables_values_id placeholder "
-                        f"{vals_placeholder!r}: {len(siblings)} rows were created under "
-                        f"variables config {parent_ulid}. Refusing to write an "
-                        "ambiguous values link."
-                    ),
-                }
-            )
-        return None
-
-    def _apply_variable_binding(
-        self,
-        client: Any,
-        *,
-        created: CreatedConfig,
-        local_data: dict[str, Any],
-        parent_ulid: str,
-        row_ulid: str | None,
-        manifest: Manifest,
-        branch_id: int | None,
-    ) -> None:
-        """PUT the resolved variables link, rewrite local, refresh manifest hashes.
-
-        ``local_data`` is the pristine on-disk ``_config.yml`` dict; a deep
-        copy is code-merged to build the full PUT body so blocks/code stay only
-        in their companion files. Uses :meth:`KeboolaClient.update_config`
-        (PUT) directly -- **not** ``set_variables``, which would create a
-        *second* variables config.
-        """
-        merged = copy.deepcopy(local_data)
-        merge_code_files(created.component_id, merged, created.config_dir)
-        _name, _description, configuration = local_config_to_api(merged)
-        configuration["variables_id"] = parent_ulid
-        if row_ulid:
-            configuration["variables_values_id"] = row_ulid
-
-        client.update_config(
-            component_id=created.component_id,
-            config_id=created.config_id,
-            configuration=configuration,
-            change_description="Resolve variables link via kbagent sync push",
-            branch_id=branch_id,
-        )
-        logger.info(
-            "Resolved variables link for %s/%s -> variables_id=%s variables_values_id=%s",
-            created.component_id,
-            created.config_id,
-            parent_ulid,
-            row_ulid,
-        )
-
-        # Rewrite the local _configuration_extra to the ULIDs (pristine data:
-        # no merged blocks leak into _config.yml).
-        extra = local_data.setdefault("_configuration_extra", {})
-        extra["variables_id"] = parent_ulid
-        if row_ulid:
-            extra["variables_values_id"] = row_ulid
-        self._write_config_file(created.config_dir, local_data)
-
-        # config_hash includes _configuration_extra, so refresh the stored
-        # hashes from the post-rewrite disk state or sync diff sees a conflict.
-        hashes = self._compute_config_hashes(created.config_dir, created.component_id)
-        target_branch = branch_id or 0
-        for cfg in manifest.configurations:
-            if (
-                cfg.branch_id == target_branch
-                and cfg.component_id == created.component_id
-                and cfg.id == created.config_id
-            ):
-                cfg.metadata["pull_hash"] = hashes.file_hash
-                cfg.metadata["pull_config_hash"] = hashes.cfg_hash
-                cfg.metadata["pull_extra_hashes"] = hashes.extra_hashes
-                break
-
-    def _resolve_flow_task_bindings(
-        self,
-        client: Any,
-        *,
-        created_configs: list[CreatedConfig],
-        created_id_map: dict[tuple[str, str], str],
-        manifest: Manifest,
-        branch_id: int | None,
-    ) -> FlowBindingResult:
-        """Remap keboola.flow task ``configId``s from source ids to ULIDs (Phase D).
-
-        A ``keboola.flow`` config runs other configs via
-        ``configuration.tasks[].task.configId`` (job-type tasks only; on disk
-        these live under ``_configuration_extra.tasks``). When a flow is created
-        in the same push as the configs it targets -- e.g. a ``sync clone`` of a
-        reference project -- those task ``configId``s still point at the source
-        config ids. This pass (mirroring the Phase-C variable backfill) resolves
-        each job task's ``(componentId, configId)`` via ``created_id_map`` to the
-        ULID assigned this push, PUTs the corrected flow, rewrites the local
-        ``_config.yml``, and refreshes the manifest hashes so a re-push is clean.
-
-        A no-op when no flow was created this push, or when no task references a
-        config created this push (the id is left untouched -- it may legitimately
-        point at a pre-existing config).
-        """
-        result = FlowBindingResult()
-        for created in created_configs:
-            if created.component_id != FLOW_COMPONENT_ID:
-                continue
-            local_data = self._read_config_file(created.config_dir)
-            if local_data is None:
-                continue
-            extra = local_data.get("_configuration_extra")
-            if not isinstance(extra, dict):
-                continue
-            tasks = extra.get("tasks")
-            if not isinstance(tasks, list):
-                continue
-
-            remapped = self._remap_flow_tasks_in_place(tasks, created_id_map)
-            if not remapped:
-                continue
-
-            try:
-                self._apply_flow_task_binding(
-                    client,
-                    created=created,
-                    local_data=local_data,
-                    manifest=manifest,
-                    branch_id=branch_id,
-                )
-            except KeboolaApiError as exc:
-                result.errors.append(
-                    {
-                        "change_type": "flow_task_link",
-                        "error_code": ErrorCode.API_ERROR,
-                        "component_id": created.component_id,
-                        "config_id": created.config_id,
-                        "message": str(exc),
-                    }
-                )
-                continue
-            result.configs_rewritten += 1
-            result.tasks_remapped += remapped
-        return result
-
-    @staticmethod
-    def _remap_flow_tasks_in_place(
-        tasks: list[Any], created_id_map: dict[tuple[str, str], str]
-    ) -> int:
-        """Rewrite job-task ``configId``s in a flow ``tasks`` list in place.
-
-        Returns the number of task references actually remapped. Only
-        ``task.type == 'job'`` entries carry a ``configId``; notification and
-        variable tasks are skipped. A task is rewritten only when
-        ``(componentId, configId)`` matches an entry created this push.
-        """
-        remapped = 0
-        for task_entry in tasks:
-            if not isinstance(task_entry, dict):
-                continue
-            task = task_entry.get("task")
-            if not isinstance(task, dict) or task.get("type") != "job":
-                continue
-            comp = task.get("componentId")
-            old_cfg = task.get("configId")
-            if not isinstance(comp, str) or not isinstance(old_cfg, (str, int)):
-                continue
-            new_id = created_id_map.get((comp, str(old_cfg)))
-            if new_id and str(old_cfg) != new_id:
-                task["configId"] = new_id
-                remapped += 1
-        return remapped
-
-    def _apply_flow_task_binding(
-        self,
-        client: Any,
-        *,
-        created: CreatedConfig,
-        local_data: dict[str, Any],
-        manifest: Manifest,
-        branch_id: int | None,
-    ) -> None:
-        """PUT a remapped flow, rewrite local ``_config.yml``, refresh hashes.
-
-        ``local_data`` already carries the remapped task ``configId``s (the
-        caller mutated ``_configuration_extra.tasks`` in place). A deep copy is
-        code-merged to build the PUT body (no-op for flows, which carry no code)
-        so the API receives the corrected ``configuration.tasks``.
-        """
-        merged = copy.deepcopy(local_data)
-        merge_code_files(created.component_id, merged, created.config_dir)
-        _name, _description, configuration = local_config_to_api(merged)
-
-        client.update_config(
-            component_id=created.component_id,
-            config_id=created.config_id,
-            configuration=configuration,
-            change_description="Remap flow task configIds via kbagent sync push",
-            branch_id=branch_id,
-        )
-        logger.info(
-            "Remapped flow task configIds for %s/%s",
-            created.component_id,
-            created.config_id,
-        )
-
-        self._write_config_file(created.config_dir, local_data)
-        hashes = self._compute_config_hashes(created.config_dir, created.component_id)
-        target_branch = branch_id or 0
-        for cfg in manifest.configurations:
-            if (
-                cfg.branch_id == target_branch
-                and cfg.component_id == created.component_id
-                and cfg.id == created.config_id
-            ):
-                cfg.metadata["pull_hash"] = hashes.file_hash
-                cfg.metadata["pull_config_hash"] = hashes.cfg_hash
-                cfg.metadata["pull_extra_hashes"] = hashes.extra_hashes
-                break
 
     def _writeback_create_config_in_manifest(
         self,
